pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langgroup"
version = "0.2.0"
description = "A multiagent system framework built on LangChain and LangGraph with supervisor-based coordination"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "William Kang", email = "willysk73@outlook.com"},
]
keywords = ["langchain", "langgraph", "multiagent", "ai", "agents", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.20",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/willysk73/langgroup"
Documentation = "https://github.com/willysk73/langgroup#readme"
Repository = "https://github.com/willysk73/langgroup"
"Bug Tracker" = "https://github.com/willysk73/langgroup/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
langgroup = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s - %(message)s"
log_cli_date_format = "%H:%M:%S"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
